pyproject.toml
[tool.poetry]
name = "true-colours-discord-bot"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"
"discord.py" = "^2.3.2"

[tool.poetry.dev-dependencies]
ruff = "^0.1.8"
mypy = "^1.7.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
